+++ src/gspits/one_dim/__init__.py
@@ -1,6 +1,11 @@
 """Routines for Bose gases in one-dimensional geometries."""
 
-from .single import HTHamiltonian
+from .single import (
+    HTHamiltonian,
+    MRHamiltonian,
+    OLHTHamiltonian,
+    plane_wave_state,
+)
 from .system import (
     ExternalPotential,
     Hamiltonian,
@@ -13,7 +18,10 @@ __all__ = [
     "ExternalPotential",
     "HTHamiltonian",
     "Hamiltonian",
+    "MRHamiltonian",
+    "OLHTHamiltonian",
     "State",
     "SupportsExternalPotential",
     "external_potential",
+    "plane_wave_state",
 ]